restructure and test
