+++ src/compas/files/gltf/__init__.py
@@ -1,0 +1,23 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from .gltf import GLTF
+from .gltf_reader import GLTFReader
+from .gltf_exporter import GLTFExporter
+from .gltf_parser import GLTFParser
+from .gltf_node import GLTFNode
+from .gltf_scene import GLTFScene
+from .data_classes import MeshData
+from .constants import DEFAULT_ROOT_NAME
+
+__all__ = [
+    'GLTF',
+    'GLTFReader',
+    'GLTFParser',
+    'GLTFExporter',
+    'GLTFScene',
+    'GLTFNode',
+    'MeshData',
+    'DEFAULT_ROOT_NAME'
+]
